grid_worker/standalone/requirements.txt
polars>=0.20.0
numpy>=1.24.0
psutil>=5.9.0
pyyaml>=6.0
requests>=2.28.0
yfinance>=0.2.0
